맵 구조 및 로직 변경

+++ Source/VoxelFortress/Private/Util/VoxelDataSturct.h
@@ -18,6 +18,6 @@ public:
     int32 TestNum;
     
     UPROPERTY(EditAnywhere,BlueprintReadWrite)
-    TArray<float> voxels;
+    TArray<int> voxels;
 
 };
